+++ Sistema/002-IMPLEMENTACAO/SISAP/src/main/java/br/edu/ifpb/monteiro/ads/sisap/entities/VisitaDomiciliarConverter.java
@@ -44,7 +44,7 @@ public class VisitaDomiciliarConverter implements Converter {
 		if (valor == null) {
 			return null;
 		}
-		Integer id = ((VisitaDomiciliar) valor).getId();
+		Long id = ((VisitaDomiciliar) valor).getId();
 
 		return (id != null) ? id.toString() : null;
 	}
